Split up methods, split up tests, add new tests, mark stage=funsor
